object level permissions test

+++ tests/Export-DbaLogin.Tests.ps1
@@ -17,49 +17,45 @@ Describe "$CommandName Integration Tests" -Tags "IntegrationTests" {
     BeforeAll {
         $DefaultExportPath = Get-DbatoolsConfigValue -FullName path.dbatoolsexport
         $AltExportPath = "$env:USERPROFILE\Documents"
-        try {
-            $random = Get-Random
-            $dbname1 = "dbatoolsci_exportdbalogin1$random"
-            $login1 = "dbatoolsci_exportdbalogin_login1$random"
-            $user1 = "dbatoolsci_exportdbalogin_user1$random"
+        $random = Get-Random
+        $dbname1 = "dbatoolsci_exportdbalogin1$random"
+        $login1 = "dbatoolsci_exportdbalogin_login1$random"
+        $user1 = "dbatoolsci_exportdbalogin_user1$random"
 
-            $dbname2 = "dbatoolsci_exportdbalogin2$random"
-            $login2 = "dbatoolsci_exportdbalogin_login2$random"
-            $user2 = "dbatoolsci_exportdbalogin_user2$random"
+        $dbname2 = "dbatoolsci_exportdbalogin2$random"
+        $login2 = "dbatoolsci_exportdbalogin_login2$random"
+        $user2 = "dbatoolsci_exportdbalogin_user2$random"
 
-            $server = Connect-DbaInstance -SqlInstance $script:instance2
-            $null = $server.Query("CREATE DATABASE [$dbname1]")
-            $null = $server.Query("CREATE LOGIN [$login1] WITH PASSWORD = 'GoodPass1234!'")
-            $server.Databases[$dbname1].ExecuteNonQuery("CREATE USER [$user1] FOR LOGIN [$login1]")
+        $server = Connect-DbaInstance -SqlInstance $script:instance2
+        $null = $server.Query("CREATE DATABASE [$dbname1]")
+        $null = $server.Query("CREATE LOGIN [$login1] WITH PASSWORD = 'GoodPass1234!'")
+        $server.Databases[$dbname1].ExecuteNonQuery("CREATE USER [$user1] FOR LOGIN [$login1]")
 
-            $null = $server.Query("CREATE DATABASE [$dbname2]")
-            $null = $server.Query("CREATE LOGIN [$login2] WITH PASSWORD = 'GoodPass1234!'")
-            $null = $server.Query("ALTER LOGIN [$login2] DISABLE")
-            $null = $server.Query("DENY CONNECT SQL TO [$login2]")
+        $null = $server.Query("CREATE DATABASE [$dbname2]")
+        $null = $server.Query("CREATE LOGIN [$login2] WITH PASSWORD = 'GoodPass1234!'")
+        $null = $server.Query("ALTER LOGIN [$login2] DISABLE")
+        $null = $server.Query("DENY CONNECT SQL TO [$login2]")
 
-            if ($server.VersionMajor -lt 11) {
-                $null = $server.Query("EXEC sys.sp_addsrvrolemember @rolename=N'dbcreator', @loginame=N'$login2'")
-            } else {
-                $null = $server.Query("ALTER SERVER ROLE [dbcreator] ADD MEMBER [$login2]")
-            }
-            $null = $server.Query("GRANT SELECT ON sys.databases TO [$login2] WITH GRANT OPTION")
-            $server.Databases[$dbname2].ExecuteNonQuery("CREATE USER [$user2] FOR LOGIN [$login2]")
-            $server.Databases[$dbname2].ExecuteNonQuery("GRANT SELECT ON sys.tables TO [$user2] WITH GRANT OPTION")
-        } catch {
-            $_
+        if ($server.VersionMajor -lt 11) {
+            $null = $server.Query("EXEC sys.sp_addsrvrolemember @rolename=N'dbcreator', @loginame=N'$login2'")
+        } else {
+            $null = $server.Query("ALTER SERVER ROLE [dbcreator] ADD MEMBER [$login2]")
         }
+        $server.Databases[$dbname2].ExecuteNonQuery("CREATE USER [$user2] FOR LOGIN [$login2]")
+        $server.Databases[$dbname2].ExecuteNonQuery("GRANT SELECT ON sys.tables TO [$user2] WITH GRANT OPTION")
+
     }
     AfterAll {
-        try {
-            Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname1 -Confirm:$false
-            Remove-DbaLogin -SqlInstance $script:instance2 -Login $login1 -Confirm:$false
+        Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname1 -Confirm:$false
+        Remove-DbaLogin -SqlInstance $script:instance2 -Login $login1 -Confirm:$false
 
-            Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname2 -Confirm:$false
-            Remove-DbaLogin -SqlInstance $script:instance2 -Login $login2 -Confirm:$false
-        } catch { }
+        Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname2 -Confirm:$false
+        Remove-DbaLogin -SqlInstance $script:instance2 -Login $login2 -Confirm:$false
         $timenow = (Get-Date -uformat "%m%d%Y%H")
         $ExportedCredential = Get-ChildItem $DefaultExportPath, $AltExportPath | Where-Object { $_.Name -match "$timenow\d{4}-login.sql|Dbatoolsci_login_CustomFile.sql" }
-        $null = Remove-Item -Path $($ExportedCredential.FullName) -ErrorAction SilentlyContinue
+        if ($ExportedCredential) {
+            $null = Remove-Item -Path $($ExportedCredential.FullName) -ErrorAction SilentlyContinue
+        }
     }
 
     Context "Executes with Exclude Parameters" {
@@ -104,9 +100,11 @@ Describe "$CommandName Integration Tests" -Tags "IntegrationTests" {
         }
         It "Should Export with object level permissions" {
             $results = Export-DbaLogin -SqlInstance $script:instance2 -Login $login2 -ObjectLevel -PassThru -WarningAction SilentlyContinue
-            $results | Should Match "$login2|$dbname2"
             $results | Should Not Match "$login1|$dbname1"
-            $results | Should Match "GRANT SELECT ON \[sys\]\.\[tables\] .*$login2.* WITH GRANT OPTION"
+            $results | Should Match "GRANT SELECT ON OBJECT::\[sys\]\.\[tables\] TO \[$user2\] WITH GRANT OPTION"
+            $results | Should Match "CREATE USER \[$user2\] FOR LOGIN \[$login2\]"
+            $results | Should Match "IF NOT EXISTS"
+            $results | Should Match "USE \[$dbname2\]"
         }
         foreach ($version in $((Get-Command $CommandName).Parameters.DestinationVersion.attributes.validvalues)) {
             It "Should Export for the SQLVersion $version" {
